Funcionalidad realizada, se muestra en pantalla del detalle el estado.
