+++ project/frontend/src/pages/detalle_lote.css
@@ -1,0 +1,95 @@
+body {
+  margin: 0;
+  padding: 0;
+}
+
+.DetalleLote {
+    display: flex;
+    flex-direction: column; /* Alinea el navbar y el cuerpo en columna */
+    min-width: 1000px; /* Asegura un ancho mínimo para evitar que el diseño se rompa en pantallas pequeñas */
+    width: 100vw; /* Asegura que ocupe todo el ancho de la ventana */
+    min-height: 100vh; /* Asegura que ocupe toda la altura de la ventana */
+    background-color: #f4f3ef;
+}
+
+.Navbar {
+    position: fixed; /* Fija el navbar en la parte superior */
+    flex-direction: 0; /* Evita que el navbar se encoja si hay poco espacio */
+}
+
+.Cuerpo {
+    flex-grow: 1; /* Permite que .Cuerpo ocupe todo el espacio restante disponible */
+    width : 100%;
+    display: flex;
+    overflow-y: auto; /* Si el contenido es muy largo, permite scroll solo en el cuerpo */
+    color: black;
+}
+
+.detalle-lote {
+  background: #ffffff;
+  border-radius: 16px;
+  padding: 20px 30px;
+  max-width: 600px;
+  margin: 40px auto;
+  box-shadow: 0 4px 15px rgba(0,0,0,0.1);
+  font-family: "Inter", sans-serif;
+}
+
+.detalle-lote h2 {
+  color: #2c3e50;
+  border-bottom: 2px solid #50b498;
+  padding-bottom: 10px;
+  margin-bottom: 20px;
+  text-align: center;
+}
+
+.detalle-info p {
+  margin: 10px 0;
+  color: #555;
+  font-size: 15px;
+}
+
+.precio-original {
+  text-decoration: line-through;
+  color: #888;
+  font-size: 16px;
+}
+
+.precio-rescate {
+  color: #50b498; 
+  font-size: 20px;
+  font-weight: 700;
+}
+
+.estado {
+  padding: 4px 10px;
+  border-radius: 8px;
+  color: white;
+  font-weight: bold;
+}
+
+/* Use a hyphenated class name for "no disponible" (apply in HTML as class="estado no-disponible") */
+.estado.publicado { background: #3498db; }
+.estado.disponible { background: #2ecc71; }
+.estado.no-disponible { background: #e67e22; }
+.estado.cancelado { background: #e74c3c; }
+.estado.reservado { background: #9b59b6; }
+.estado.oculto { background: #7f8c8d; }
+
+.detalle-lote button {
+    padding: 10px;
+    border: none;
+    background: #017007;
+    border-radius: 5px;
+    cursor: pointer;
+    font-weight: bold;
+    width: 48%;
+    margin: 5px auto;
+    transition: background 0.2s ease;
+    display: block;
+}
+
+.detalle-lote button:hover {
+    background: #01b50a;
+    color: black;
+}
